Moved hyper support in a separate module. One can now use the hyper support without necessarily using our server, nor tokio net.
