+++ Cargo.toml
@@ -8,14 +8,15 @@ repository = "https://github.com/restatedev/sdk-rust"
 
 [features]
 default = ["http_server", "rand", "uuid"]
-http_server = ["hyper", "http-body-util", "hyper-util", "tokio/net", "tokio/signal", "restate-sdk-shared-core/http"]
+hyper = ["dep:hyper", "http-body-util", "restate-sdk-shared-core/http"]
+http_server = ["hyper", "hyper/server", "hyper/http2", "hyper-util", "tokio/net", "tokio/signal"]
 
 [dependencies]
 bytes = "1.6.1"
 futures = "0.3"
 http = "1.1.0"
 http-body-util = { version = "0.1", optional = true }
-hyper = { version = "1.4.1", optional = true, features = ["server", "http2"] }
+hyper = { version = "1.4.1", optional = true}
 hyper-util = { version = "0.1", features = ["tokio", "server", "server-graceful", "http2"], optional = true }
 pin-project-lite = "0.2"
 rand = { version = "0.8.5", optional = true }
@@ -25,7 +26,7 @@ restate-sdk-shared-core = { version = "0.0.5" }
 serde = "1.0"
 serde_json = "1.0"
 thiserror = "1.0.63"
-tokio = { version = "1", default-features = false, features = ["sync", "macros"] }
+tokio = { version = "1", default-features = false, features = ["sync"] }
 tower-service = "0.3"
 tracing = "0.1"
 uuid = { version = "1.10.0", optional = true }
