+++ examples/greeter.rs
@@ -16,7 +16,7 @@ impl Greeter for GreeterImpl {
 #[tokio::main]
 async fn main() {
     tracing_subscriber::fmt::init();
-    HyperServer::new(
+    HttpServer::new(
         Endpoint::builder()
             .with_service(GreeterImpl.serve())
             .build(),
