+++ examples/run.rs
@@ -33,7 +33,7 @@ impl RunExample for RunExampleImpl {
 #[tokio::main]
 async fn main() {
     tracing_subscriber::fmt::init();
-    HyperServer::new(
+    HttpServer::new(
         Endpoint::builder()
             .with_service(RunExampleImpl(reqwest::Client::new()).serve())
             .build(),
